Task1 and 2
Moving models and simulation runing
