+++ labothappy/toolbox/spliter/spliter.py
@@ -4,6 +4,19 @@ Created on Fri May 10 14:31:24 2024
 
 @author: Basile
 """
+import sys
+import os
+
+# Get the absolute path of the directory that contains the script (simulation_model.py)
+current_dir = os.path.dirname(os.path.abspath(__file__))
+
+# Determine the project root directory (which contains both 'connector' and 'component')
+project_root = os.path.abspath(os.path.join(current_dir, '..', '..', '..'))
+
+# Add the project root to sys.path if it's not already there
+if project_root not in sys.path:
+    sys.path.insert(0, project_root)
+###########################################
 
 from connector.mass_connector import MassConnector
 from CoolProp.CoolProp import PropsSI
